Support optional use of remote filename for output
As an optional feature, if useremotefilename is set to 1, then the local
filename will use the filename found in the enclosure.

This fixes some problems with bashpodder I experienced when trying to
download a back catalog of a show.  With this setting, one can retrieve
many old episodes into a given date's directory.